Move biometric security code from BrowserActivity to DuckDuckGoActivity
