+++ app/build.gradle
@@ -22,8 +22,8 @@ android {
         minSdk min_sdk
         targetSdk target_sdk
         compileSdk compile_sdk
-        versionCode 202505072
-        versionName "1.3.3"
+        versionCode 202505131
+        versionName "1.3.5"
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         archivesBaseName = "kahf_browser_${versionName}_${versionCode}"
         vectorDrawables.useSupportLibrary = true
@@ -391,7 +391,6 @@ dependencies {
     implementation JakeWharton.timber
     implementation 'nl.dionsegijn:konfetti:1.2.5'
     implementation 'com.github.KahfAds:adsdk:2.2.0'
-    implementation 'androidx.biometric:biometric:1.1.0'
 
     // App update
     implementation 'com.google.android.play:app-update:2.1.0'
