+++ common/common-ui/src/main/java/com/duckduckgo/common/ui/DuckDuckGoActivity.kt
@@ -28,6 +28,9 @@ import android.view.MenuItem
 import android.view.View
 import android.view.WindowManager.LayoutParams
 import androidx.appcompat.widget.Toolbar
+import androidx.biometric.BiometricManager
+import androidx.biometric.BiometricPrompt
+import androidx.core.content.ContextCompat
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import androidx.localbroadcastmanager.content.LocalBroadcastManager
@@ -46,6 +49,17 @@ abstract class DuckDuckGoActivity : DaggerActivity() {
     @Inject lateinit var themingDataStore: ThemingDataStore
 
     private var themeChangeReceiver: BroadcastReceiver? = null
+
+    private lateinit var biometricPrompt: BiometricPrompt
+    private val promptInfo: BiometricPrompt.PromptInfo by lazy {
+        BiometricPrompt.PromptInfo.Builder()
+            .setTitle(getString(R.string.kahf_safegaze_unlock))
+            .setAllowedAuthenticators(
+                BiometricManager.Authenticators.BIOMETRIC_WEAK or
+                    BiometricManager.Authenticators.DEVICE_CREDENTIAL,
+            )
+            .build()
+    }
 
     @SuppressLint("MissingSuperCall")
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -143,6 +157,51 @@ abstract class DuckDuckGoActivity : DaggerActivity() {
         }
         themingDataStore.theme = newTheme
         sendThemeChangedBroadcast()
+    }
+
+    fun isAnySecurityEnabled(): Boolean {
+        val biometricManager = BiometricManager.from(this@DuckDuckGoActivity)
+        return biometricManager.canAuthenticate(
+            BiometricManager.Authenticators.BIOMETRIC_WEAK or
+                BiometricManager.Authenticators.DEVICE_CREDENTIAL,
+        ) == BiometricManager.BIOMETRIC_SUCCESS
+    }
+
+    fun showBiometricPrompt(function: (authenticated: Boolean, msgId: Int) -> Unit) {
+        biometricPrompt = BiometricPrompt(this, ContextCompat.getMainExecutor(this),
+            object : BiometricPrompt.AuthenticationCallback() {
+                override fun onAuthenticationSucceeded(result: BiometricPrompt.AuthenticationResult) {
+                    super.onAuthenticationSucceeded(result)
+                    function.invoke(true, 0)
+                }
+
+                override fun onAuthenticationError(errorCode: Int, errString: CharSequence) {
+                    super.onAuthenticationError(errorCode, errString)
+
+                    val messageResId = when (errorCode) {
+                        BiometricPrompt.ERROR_NO_DEVICE_CREDENTIAL ->
+                            R.string.kahf_no_device_credential
+                        BiometricPrompt.ERROR_NO_BIOMETRICS ->
+                            R.string.kahf_no_biometrics
+                        BiometricPrompt.ERROR_HW_NOT_PRESENT ->
+                            R.string.kahf_no_biometric_hardware
+                        BiometricPrompt.ERROR_HW_UNAVAILABLE ->
+                            R.string.kahf_biometric_unavailable
+                        BiometricPrompt.ERROR_LOCKOUT ->
+                            R.string.kahf_too_many_attempts
+                        else -> R.string.kahf_authentication_error
+                    }
+
+                    function.invoke(false, messageResId)
+                }
+
+                override fun onAuthenticationFailed() {
+                    super.onAuthenticationFailed()
+                    function.invoke(false, R.string.kahf_not_recognized)
+                }
+            })
+
+        biometricPrompt.authenticate(promptInfo)
     }
 
     protected inline fun <reified V : ViewModel> bindViewModel() = lazy {
